requirements.txt
wheel
pyyaml
pyside6
GitPython
rsa
ipython
pympler
dataclassy
pyzmq
blosc2
torch
torchvision
torchaudio